Move read_block test
