+++ distributed/tests/test_s3.py
@@ -9,7 +9,7 @@ from dask.imperative import Value
 from distributed import Executor
 from distributed.executor import _wait, Future
 from distributed.s3 import (read_bytes, read_text,
-        read_block, seek_delimiter, S3FileSystem, _read_text)
+        seek_delimiter, S3FileSystem, _read_text)
 from distributed.utils import get_ip
 from distributed.utils_test import gen_cluster, loop, cluster, slow
 
@@ -158,28 +158,6 @@ def test_read_bytes_block(s, a, b):
         testlines = b"".join(files.values()).split(b'\n')
         assert set(ourlines) == set(testlines)
     yield e._shutdown()
-
-
-def test_read_block():
-    delimiter = b'\n'
-    data = delimiter.join([b'123', b'456', b'789'])
-    f = io.BytesIO(data)
-
-    assert read_block(f, 1, 2) == b'23'
-    assert read_block(f, 0, 1, delimiter=b'\n') == b'123\n'
-    assert read_block(f, 0, 2, delimiter=b'\n') == b'123\n'
-    assert read_block(f, 0, 3, delimiter=b'\n') == b'123\n'
-    assert read_block(f, 0, 5, delimiter=b'\n') == b'123\n456\n'
-    assert read_block(f, 0, 8, delimiter=b'\n') == b'123\n456\n789'
-    assert read_block(f, 0, 100, delimiter=b'\n') == b'123\n456\n789'
-    assert read_block(f, 1, 1, delimiter=b'\n') == b''
-    assert read_block(f, 1, 5, delimiter=b'\n') == b'456\n'
-    assert read_block(f, 1, 8, delimiter=b'\n') == b'456\n789'
-
-    for ols in [[(0, 3), (3, 3), (6, 3), (9, 2)],
-                [(0, 4), (4, 4), (8, 4)]]:
-        out = [read_block(f, o, l, b'\n') for o, l in ols]
-        assert b"".join(filter(None, out)) == data
 
 
 @gen_cluster(timeout=60)
